Added playlists as a filterable property. Should solve #128
Combined with the ability to randomly cycle through the currently filtered wallpaper list, this should give the ability to cycle through wallpapers like in the Windows app.

This also means you can cycle through the union of two or more playlists or some other combination of playlists and other filters, which is a pretty cool enhancement on the Windows app's capabilities.

Needed to shore up some of the async code as part of the solution to a race condition when configuring multiple displays to shuffle the filter simultaneously.