Refactor audio mute handling in MpvObject to use audio track ID
Previously, muting/unmuting would not unload the entire audio backend, which would
cause the audio track to still be loaded and consume some resources. Now, we set the
audio track ID to "no" when muting, which completely disables audio processing in mpv.
We probably won't notice a difference in most cases, but it was bugging me.
It was setting an audio track at 48000 Hz even when muted and it was not
allowing Pipewire to change frequencies on the fly.
Not sure how to also handle it for the other video backend.